split up working_dir test file that was timing out
